+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to `libcasm-monte` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added `AtomInfo` and `AtomInfoMap` to `libcasm.monte.events`.
+
+### Fixed
+
+- Fixed documentation and binding errors in `libcasm.monte.events` and `libcasm.monte.sampling`.
+
+
 ## [2.0a4] - 2024-12-20
 
 ### Fixed